Lighter compilation cache
Slim down the compilation cache.  Before, the cache was sized for two
words per entry, plus six words per occupied entry (a size-4 FixedData
key).  In the new scheme, there is just one word in the entry, and it
just holds a SharedFunctionInfo.

The trade-off that was made was to add a word to all SharedFunctionInfo
instances, to hold a pointer to the "parent" SharedFunctionInfo.  This
lets us match a string + Context with an already compiled
SharedFunctionInfo.

This commit also simplifies Script::source() -- it will always be a
string.  There is no part of V8 that ever stores a non-string in that
slot in a way that is visible to any part of the code.  As such, change
its type to String*, and adapt associated logic.

This commit does not increase the number of instructions executed as d8
reaches its prompt, nor does it impact startup memory usage.  I don't
have a chromium build handy to test its impact on larger workloads.

TEST=no additional tests; passes test suite in release and debug mode