Scripts compress their sources
Add a new CompressedSource object that is a compressed representation of
a script's source code.  The codec has been made in such a way that it
is easy to hash, compare, and extract substrings without decompressing
the whole script.  In this way we can lower the memory footprint of
source code while still keeping it around.

The codec uses a fixed dictionary bytecode scheme derived from a corpus
of data.  See https://gitorious.org/corps/corps/blobs/master/README for
more details.  The generated codec files code-table.inc.c,
decoder.inc.c, and encoder.inc.c are exactly the ones that are present
in the corps source tree, in
https://gitorious.org/corps/corps/trees/master/corps/js.

This patch does have some overhead: compressing and decompressing
strings is not free.

The overall impact of this patch appears to be that it causes 20%
slowdowns on startup (running d8, about 10M instructions), both in
memory (due to the need to keep both compressed and raw data in memory)
and in CPU (due to the need to compress).  It is hoped that the
long-term memory impact of the patch will pay for its costs, but that is
yet to be determined.

As an optimization, the CompressedSource object keeps a weak reference
on the original source string, cleared before the next major garbage
collection.  This lowers the cost of decompression soon after the script
was made.